refactor: use a single `UnboundedExec` for testing
